Added a FIXME comment to LAMbS_mo.c releated to the mo transition notifier mechanism. Changed the LAMbS_current_moi declaration from u64 to s32 in the header LAMbS_mo.h, which is the correct declaration. This mistake was causing kernel panics during the insmod of pbs_module.

+++ governor/LAMbS_mo.h
@@ -4,7 +4,7 @@
 #include "LAMbS_molookup.h"
 #include "LAMbS_mostat.h"
 
-extern u64  LAMbS_current_moi;
+extern s32  LAMbS_current_moi;
 
 int     LAMbS_mo_init(int verbose);
 void    LAMbS_mo_uninit(void);
